requirements.txt
django-tex
jinja2-django-tags
pytest
pytest-django
pytest-mock
factory_boy
